test(desktop): split git repository specs
