+++ scripts/check-file-sizes.mjs
@@ -70,7 +70,6 @@ const ALLOWLIST = new Set([
   "apps/desktop/src/__tests__/components/timeline/AgentTreeView.test.ts",
   "apps/desktop/src/__tests__/composables/useCachedFetch.test.ts",
   "apps/desktop/src/__tests__/composables/useImportFlow.test.ts",
-  "apps/desktop/src/composables/__tests__/useGitRepository.test.ts",
   "apps/desktop/src/__tests__/views/analytics-views.test.ts",
 ]);
